requirements.txt
asttokens==2.0.5
backcall==0.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
charset-normalizer==2.1.1
cycler==0.11.0
dataclasses==0.6
debugpy==1.6.0
decorator==5.1.1
executing==0.8.3
fonttools==4.37.1
idna==3.4
ipykernel==6.13.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.10.0
kiwisolver==1.4.4
matplotlib==3.5.3
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.4
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.29
psutil==5.9.1
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.1.0
requests==2.28.1
scipy==1.9.1
seaborn==0.12.0
six==1.16.0
stack-data==0.2.0
thop==0.1.1.post2209072238
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tornado==6.1
tqdm==4.64.1
traitlets==5.2.2.post1
typing-extensions==3.10.0.0
urllib3==1.26.12
wcwidth==0.2.5
